requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==24.2.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cssselect2==0.7.0
Django==5.0.2
django-auditlog==3.0.0
django-cors-headers==4.3.1
django-crum==0.7.9
django-minio-storage==0.5.7
django-seeding==1.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
drf-spectacular-sidecar==2024.1.1
fonttools==4.54.1
h11==0.14.0
html5lib==1.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
minio==7.2.7
numpy==1.26.4
pandas==2.2.2
pillow==10.2.0
psycopg==3.1.18
psycopg-binary==3.1.18
pycparser==2.22
pycryptodome==3.20.0
pydyf==0.11.0
PyJWT==2.8.0
pyphen==0.16.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.0
tinycss2==1.3.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.6
weasyprint==62.3
webencodings==0.5.1
zopfli==0.2.3
